@@ -1,14 +1,17 @@
 [build-system]
 requires = [
 	"ase",
+	"atomistics",
 	"matgl",
 	"matplotlib",
+	"nglview",
 	"numpy",
 	"pandas",
 	"phonopy",
 	"pyiron_atomistics",
 	"pyiron_base",
 	"pyiron_contrib",
+	"pyiron_workflow",
 	"pymatgen",
 	"scipy",
 	"sqlalchemy",
@@ -23,33 +26,34 @@ name = "pyiron_nodes"
 description = "pyiron_nodes - Nodes for `pyiron_workflows` with a focus on atomistic workflows."
 readme = "docs/README.md"
 keywords = [ "pyiron",]
-requires-python = ">=3.10, <3.13"
+requires-python = ">=3.11, <3.13"
 classifiers = [
     "Development Status :: 3 - Alpha",
     "Topic :: Scientific/Engineering",
     "License :: OSI Approved :: BSD License",
     "Intended Audience :: Science/Research",
     "Operating System :: Unix",  # Because of Lammps dependency -- can change in the future
-    "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
 ]
 dependencies = [
     "ase==3.23.0",
-	"atomistics==0.1.31",
-	"matgl==1.1.2",
-	"matplotlib==3.9.1",
+	"atomistics==0.1.32",
+	"lammps==2024.06.27",
+	"matgl==1.1.3",
+	"matplotlib==3.9.2",
 	"nglview==3.1.2",
 	"numpy==1.26.4",
 	"pandas==2.2.2",
 	"phonopy==2.26.6",
-	"pyiron_atomistics==0.6.9",
-	"pyiron_base==0.9.10",
-	"pyiron_contrib==0.1.17",
-  	"pyiron_workflow==0.9.4",
-	"pymatgen==2024.7.18",
-	"scipy==1.14.0",
-	"sqlalchemy==2.0.31",
-	"structuretoolkit==0.0.27",
+	"pyiron_atomistics==0.6.11",
+	"pyiron_base==0.9.12",
+	"pyiron_contrib==0.1.18",
+  	"pyiron_workflow==0.10.0",
+	"pymatgen==2024.8.9",
+	"scipy==1.14.1",
+	"sqlalchemy==2.0.32",
+	"structuretoolkit==0.0.28",
 ]
 dynamic = [ "version",]
 authors = [
