@@ -1,13 +1,12 @@
+from __future__ import annotations
+
+from ase import Atoms
 from pyiron_workflow import as_function_node
 from typing import Optional, Union
 
-# Huge savings when replacing pyiron_atomistics atoms class with ase one!! (> 5s vs 40 ms)
-# from pyiron_atomistics.atomistics.structure.atoms import Atoms
-from ase import Atoms
-
 
 @as_function_node("structure")
-def repeat(structure: Optional[Atoms], repeat_scalar: int = 1) -> Atoms:
+def repeat(structure: Atoms, repeat_scalar: int = 1) -> Atoms:
     return structure.repeat(repeat_scalar)
 
 
@@ -22,8 +21,8 @@ def apply_strain(
     return struct
 
 
-@as_function_node()
-def create_vacancy(structure, index: Optional[int] = 0):
+@as_function_node
+def create_vacancy(structure, index: int | None = 0):
     structure = structure.copy()
     if index is not None:
         del structure[index]
@@ -58,11 +57,3 @@ def rotate_axis_angle(
     structure_rotated = structure.copy()
     structure_rotated.rotate(a=angle, v=axis, center=center, rotate_cell=rotate_cell)
     return structure_rotated
-
-
-nodes = [
-    repeat,
-    apply_strain,
-    create_vacancy,
-    rotate_axis_angle,
-]
