@@ -25,7 +25,7 @@ def PlotDataFrame(
 
 @as_function_node("fig")
 def Scatter(
-        x: Optional[list | np.ndarray] = None, y: Optional[list | np.ndarray] = None
+    x: Optional[list | np.ndarray] = None, y: Optional[list | np.ndarray] = None
 ):
     from matplotlib import pyplot as plt
 
@@ -51,13 +51,13 @@ def Histogram(x: Optional[list | np.ndarray], bins: int = 50):
 
 @as_function_node("axis")
 def Plot(
-        x: Optional[list | np.ndarray | pandas.core.series.Series] = None,
-        y: Optional[list | np.ndarray | pandas.core.series.Series] = None,
-        axis: Optional[object] = None,
-        title: Optional[str] = "",
-        color: Optional[str] = "b",
-        symbol: Optional[str] = "o",
-        legend_label: Optional[str] = "",
+    x: Optional[list | np.ndarray | pandas.core.series.Series] = None,
+    y: Optional[list | np.ndarray | pandas.core.series.Series] = None,
+    axis: Optional[object] = None,
+    title: Optional[str] = "",
+    color: Optional[str] = "b",
+    symbol: Optional[str] = "o",
+    legend_label: Optional[str] = "",
 ):
     from matplotlib import pyplot as plt
 
@@ -73,10 +73,10 @@ def Plot(
 
 @as_function_node("linspace")
 def Linspace(
-        start: Optional[int | float] = 0.0,
-        stop: Optional[int | float] = 1.0,
-        num: Optional[int] = 50,
-        endpoint: Optional[bool] = True,
+    start: Optional[int | float] = 0.0,
+    stop: Optional[int | float] = 1.0,
+    num: Optional[int] = 50,
+    endpoint: Optional[bool] = True,
 ):
     from numpy import linspace
 
@@ -90,10 +90,10 @@ def Mean(numbers: Optional[list | np.ndarray] = None):
 
 @as_function_node("axes")
 def Subplot(
-        nrows: Optional[int] = 1,
-        ncols: Optional[int] = 1,
-        sharex: Optional[bool] = False,
-        sharey: Optional[bool] = False,
+    nrows: Optional[int] = 1,
+    ncols: Optional[int] = 1,
+    sharex: Optional[bool] = False,
+    sharey: Optional[bool] = False,
 ):
     from matplotlib import pyplot as plt
 
@@ -109,16 +109,3 @@ def Title(axis: Optional[object] = None, title: Optional[str] = ""):
         axis = plt
 
     return axis.set_title(title)
-
-
-nodes = [
-    PlotDataFrame,
-    Scatter,
-    Plot,
-    Histogram,
-    ShowArray,
-    Linspace,
-    Mean,
-    Subplot,
-    Title
-]
