@@ -1,23 +1,17 @@
 from __future__ import annotations
 
-
 from typing import Optional
 
-# from pyiron_atomistics.atomistics.structure.atoms import Atoms
-
+from pyiron_atomistics.lammps.control import LammpsControl
 from pyiron_workflow import as_function_node
-from pyiron_workflow.workflow import Workflow
 
 from pyiron_nodes.atomistic.calculator.data import (
     InputCalcMinimize,
     InputCalcMD,
     InputCalcStatic,
 )
-
 from pyiron_nodes.dev_tools import FileObject, parse_input_kwargs
 from pyiron_nodes.dev_tools import wf_data_class
-
-from pyiron_atomistics.lammps.control import LammpsControl
 
 
 @as_function_node("calculator")
@@ -171,7 +165,7 @@ class GenericOutput:
     forces = []
 
 
-@as_function_node()
+@as_function_node
 def Collect(
     out_dump,
     out_log,
