@@ -1,9 +1,9 @@
-from pyiron_workflow import as_function_node
-from typing import Optional, Union
+from __future__ import annotations
 
-# Huge savings when replacing pyiron_atomistics atoms class with ase one!! (> 5s vs 40 ms)
-# from pyiron_atomistics.atomistics.structure.atoms import Atoms
+from typing import Optional
+
 from ase import Atoms
+from pyiron_workflow import as_function_node
 
 
 @as_function_node("volume")
@@ -14,6 +14,3 @@ def volume(structure: Optional[Atoms] = None) -> float:
 @as_function_node("number_of_atoms")
 def number_of_atoms(structure: Optional[Atoms] = None) -> int:
     return structure.get_number_of_atoms()
-
-
-nodes = [volume, number_of_atoms]
